+++ go.mod
@@ -1,1 +1,3 @@
 module github.com/GritsyukLeonid/pastebin-go
+
+go 1.23.5
