+++ internal/model/shorturl.go
@@ -1,0 +1,14 @@
+package model
+
+type ShortURL struct {
+	original string
+	hash     string
+}
+
+func NewShortURL(original string) *ShortURL {
+	return &ShortURL{original: original}
+}
+
+func (s *ShortURL) Original() string { return s.original }
+func (s *ShortURL) SetHash(h string) { s.hash = h }
+func (s *ShortURL) Hash() string     { return s.hash }
